refactor(hub): extract refs/links enforcement to with-shape/links/vault-facade (Phase 5 A5)
