+++ scripts/check-architecture.mjs
@@ -605,7 +605,8 @@ const KERNEL_SURFACE_BUDGET = {
   // Bumped 4547→4559 (bundle includes blobs): dump() now enumerates the blob collections (global _blob_index/_blob_chunks/_blob_eviction_audit + per-collection _blob_slots_*/_blob_versions_*) so blob "covers" travel in the .noydb bundle; +12 lines (inlined name literals + computed internalNames array + explanatory comment), no blob runtime pulled into the hot path.
   // Lowered 4559→4410 (Phase 5 A3: periods extraction): the close/open/list/get methods + `_assertTsWritable` guard + `_loadPeriodsCache`/`_writePeriodRecord`/`_decryptPeriodRecord` helpers + the `periodCache` field moved to `with-audit/periods/vault-facade.ts` (`VaultPeriods`); vault.ts holds a facade instance + thin delegators (`_assertTsWritable` still the gate-bus entry point).
   // Lowered 4410→4135 (Phase 5 A4: backup extraction): `dump`/`load`/`verifyBackupIntegrity`/`exportJSON` (incl. the blob-collection enumeration in dump() and the branchy chain+data-envelope integrity walk) moved to `vault-backup.ts`; vault.ts holds thin delegators + a `backupContext()` builder exposing the read paths and the post-load mutation seams (reloadKeyring/cache-clear/ledger-reset).
-  'packages/hub/src/vault.ts': 4135,
+  // Lowered 4135→3824 (Phase 5 A5: refs/links enforcement extraction): `enforceRefsOnPut`/`enforceRefsOnDelete`/`enforceLinksOnDelete`/`resolveRef`/`resolveSource`/`checkIntegrity` bodies + the `cascadeInProgress` cycle-breaker set moved to `with-shape/links/vault-facade.ts` (`VaultLinks`); vault.ts holds a facade instance + thin delegators (the `refEnforcer`/`joinResolver` ctor seam is unchanged — Collection still passes `this`). The ref/link registries stay vault-resident (populated by collection()/link(), read by the backup path) and arrive by reference.
+  'packages/hub/src/vault.ts': 3825,
   // Bumped 2920 → 2960 (2026-06): two genuinely-core additions landed —
   // #313's `openVault` no-self-provision pre-gate (a 1-line call; the policy
   // logic itself was extracted to team/keyring.ts as `assertKeyringOpenAllowed`),
